% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casthex.R
\name{casthex}
\alias{casthex}
\title{Extract HEX codes from image url}
\usage{
casthex(x)
}
\arguments{
\item{infile}{Path to the input urls}
}
\value{
hex codes
}
\description{
List as many URLs to image files as you want
}
