% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casthex.R
\name{getHexPrimaries}
\alias{getHexPrimaries}
\title{Load URL}
\usage{
getHexPrimaries(img, pcnt.threshold = 0.05)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
hex codes
}
\description{
load image URL and extract HEX codes
}
